+++ pages/contract-verification.tsx
@@ -4,12 +4,12 @@ import React from 'react';
 import type { Props } from 'nextjs/getServerSideProps';
 import PageNextJs from 'nextjs/PageNextJs';
 
-import ContractVerificationForAddress from 'ui/pages/ContractVerificationForAddress';
+import ContractVerification from 'ui/pages/ContractVerification';
 
 const Page: NextPage<Props> = (props: Props) => {
   return (
-    <PageNextJs pathname="/address/[hash]/contract-verification" query={ props }>
-      <ContractVerificationForAddress/>
+    <PageNextJs pathname="/contract-verification" query={ props }>
+      <ContractVerification/>
     </PageNextJs>
   );
 };
